Delete modal added. Delete all and restore all functionality added.

+++ app/Http/Controllers/BannerController.php
@@ -84,7 +84,9 @@ class BannerController extends Controller
 
     public function getDelete($banner_id){
         $banner = Banner::findOrFail($banner_id);
-        unlink(public_path() . "\\banner\\" . $banner->image);
+        if(!empty($banner->image)) {
+            unlink(public_path() . "\\banner\\" . $banner->image);
+        }
         $banner->delete();
         return redirect()->route('backend.banner.delete.page')->with(['success' => 'Successfully deleted']);
     }
@@ -113,5 +115,26 @@ class BannerController extends Controller
         $banner->status = "published";
         $banner->update();
         return redirect()->route('backend.banner');
+    }
+
+    public function DeleteAll(){
+        $banners = Banner::all();
+        foreach($banners as $banner){
+            if($banner->status = "trash"){
+                $banner->delete();
+            }
+        }
+        return redirect()->route('backend.banner.delete.banner')->with(['success' => 'Trash Cleared']);
+    }
+
+    public function RestoreAll(){
+        $banners = Banner::all();
+        foreach($banners as $banner){
+            if($banner->status = "trash"){
+                $banner->status = "published";
+                $banner->update();
+            }
+        }
+        return redirect()->route('backend.banner')->with(['success' => 'All items restored']);
     }
 }
